// src/types/metadata.ts
export interface Metadata {
  totalPlayCount: number;
  totalPlayerCount: number;
  totalHighestScore: number;
  totalHighestScorePlayer: string;
  totalLowestScore: number;
  totalLowestScorePlayer: string;
  totalMostPlayedPlayer: string;
  totalMostPlayedPlayerCount: number;
  totalMostPlayedDate: Date;
  totalMostPlayedDateCount: number;
}